engine: address Phase 5 wasm-backend review feedback
The wasm backend fully unrolls every array reducer and BeginIter/
BeginBroadcastIter loop at compile time, turning each element address
into a wasm constant. There was no upper bound on that unrolling, so a
large arrayed model -- especially nested iterations whose counts
multiply -- could emit a function body exceeding what wasm engines
accept. Add a cumulative per-function unroll budget on EmitState
(MAX_UNROLL_UNITS = 65_536, the natural ceiling of a single u16
dimension), charged before each unroll site so an over-budget model is
rejected with WasmGenError::Unsupported -- a clean fall back to the VM --
rather than materializing the oversized function. The cap leaves >60x
headroom over realistic SD arrays (the corpus's largest dimension is 9)
and the corpus-wide wasm parity test still passes unchanged.

Also fold in the smaller review items: soften the Stddev self-multiply
comment (powf is libm-dependent, so x*x is not guaranteed bit-identical
across platforms, only within tolerance); drop the dead duplicated
return value from resolve_dim_list_raw and have the lone caller build
its own zero dim_ids; and replace the guarded .unwrap() pair in
emit_view_subscript_dynamic with an explicit match whose mixed-Some arm
is unreachable!, making the "both locals set together" invariant
explicit.

Tests: an over-budget reducer and an over-budget BeginIter each return
Unsupported without emitting a body; ArraySize over a huge view stays
free (no element reads); and a real runtime-OOB ViewSubscriptDynamic
feeding every reducer yields NaN end-to-end, with the in-range
counterpart matching the VM. Adding these tests pushed lower.rs past the
6000-line project lint, so the #[cfg(test)] mod tests body moves to a
sibling lower_tests.rs included via #[path] (keeping use super::* access
to the lowering module's private items); no production code moved.
